backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
python-dotenv
pytest
httpx
